move IconContainer
