+++ src/components/NftFailureFallback/NftFailureFallback.tsx
@@ -2,7 +2,7 @@ import { MouseEventHandler, useCallback, useEffect, useState } from 'react';
 import styled from 'styled-components';
 
 import colors from '~/components/core/colors';
-import IconContainer from '~/components/core/Markdown/IconContainer';
+import IconContainer from '~/components/core/IconContainer';
 import { VStack } from '~/components/core/Spacer/Stack';
 import { BaseM } from '~/components/core/Text/Text';
 import Tooltip from '~/components/Tooltip/Tooltip';
