+++ src/contexts/snow/SnowToggleIcon.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import IconContainer from '~/components/core/Markdown/IconContainer';
+import IconContainer from '~/components/core/IconContainer';
 import SnowflakeIcon from '~/icons/SnowflakeIcon';
 
 import { useSnowContext } from './SnowContext';
